+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ml).
 
 ## testing
   [1.0.10-beta]
+- [addend] #166 new action recycling and reuse
 
 ## [1.0.10-beta]