+++ CHANGELOG.md
@@ -14,6 +14,7 @@ ml).
 
 ## [1.0.10-beta]
 - [bugfix] #168 can to do a trade without devices.
+- [addend] #167 new actions of status devices: use, recycling, refurbish and management.
 
 ## [1.0.9-beta]
 - [addend] #159 external document as proof of erase of disk